Commit bad PID code
Commit some basic PID speed control code.

Why doesn't this code work? I believe that the Vex MC29 ESC's are not
appropriate motor controllers. They use a servo pulse, so that means the
speed can only be updated once every 20mS!!! That's a huge amount of
time. Next rev will attempt the same control, but using an hbridge.

Note: I disassembled the vex393 motor/encoder and its a gear with 4 bits
coupled to the dc motor with (guess) 1:10 reduction. The backlash
between the encoder and motor is negligable (one gear coupling) compared
to the resolution (4 bits).