commit super rough wall following robot code
Commit super rough wall following robot code. The code is PID based, but
not gain tuning was done. The right motor speed is controlled by the
PID/Ultrasonic sensor. The left motor spoeed is set arbitrarily. This
would all be done way nicer and smoother, this code is just a rough
proof of concept.