UPDATE -  Master branch from everyone's private branch
I updated the master branch from everyone else's branches. This Branch
at this moment is fully up to date. Please do not modify files inside
the MSE2202b Final Project CAD\Vex Components folder. That folder is a
library of standard vex components. If you need to modify a part, you
need to do a save-as and create a new part.